migrations + new features from API
Just added nullable function to some of our db tables in migrations, also now we can get good quality backdrop pics from another API using our API to get the movies and getting the IMDB ID at the same time then using IMDB ID in new API request to find a movie with that id and to get the backdrop of that movie. simple as that.